% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiEvaluate.R
\name{and_function}
\alias{and_function}
\title{and_function}
\usage{
and_function(method = EnumEvalMethod, and = NULL)
}
\arguments{
\item{method}{One of: \code{"set"} (default), \code{"prob"}, \code{"fuzzy"} or \code{"fuzzynorm"}.}

\item{and}{Some conjunctive aggregation function of the form \verb{function(num_vector)}, or \code{NULL}.}
}
\value{
Returns the function \code{and} if not \code{NULL}.
Otherwise, it determines the result depending on \code{method}:
\describe{
\item{\code{"set"}:}{\code{function(x) 0}}
\item{\code{"prob"}:}{\code{\link[base]{prod}}}
\item{\code{"fuzzy"}:}{\code{\link[base]{min}}}
\item{\code{"fuzzynorm"}:}{\code{\link[base]{min}}}
}
Fails with an error if the result is not an R function.
}
\description{
Determine the function to be used in the conjunctive aggregation step of \code{\link[=evaluate]{evaluate()}}.
}
\seealso{
\code{\link[DEXiR]{evaluate}}, \code{\link[=or_function]{or_function()}}.
}
