% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiAttributes.R
\name{att_names}
\alias{att_names}
\title{att_names}
\usage{
att_names(atts, use_id = TRUE)
}
\arguments{
\item{atts}{A vector of \link{DexiAttribute}s.}

\item{use_id}{Determines whether to return attribute IDs or original DEXi names.}
}
\value{
A character vector of attribute IDs or names.
}
\description{
Return names or IDs of \link{DexiAttribute} objects.
}
