% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiScales.R
\name{default_quality}
\alias{default_quality}
\title{default_quality}
\usage{
default_quality(order = EnumOrder, nvals)
}
\arguments{
\item{order}{`character(1)1, one of "ascending", "descending" or "none".}

\item{nvals}{\code{integer(1)}. The number of qualitative values of considered \link{DexiDiscreteScale}.}
}
\value{
character vector of length \code{nvals}, containing
\code{"bad"}, \code{"none"} or \code{"good"}.
}
\description{
Make a default discrete scale quality vector depending on the scale's \code{order} and \code{nvals}.
}
\examples{
default_quality("ascending", 5)
default_quality("descending", 5)
default_quality("none", 5)
default_quality("ascending", 2)
default_quality("ascending", 1)

}
