% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiValues.R
\name{dexi_bool}
\alias{dexi_bool}
\title{dexi_bool}
\usage{
dexi_bool(x)
}
\arguments{
\item{x}{character(1).}
}
\value{
\code{logical(1)}.
}
\description{
Convert a DEXi string to logical. \code{"TRUE"}, \code{"T"} and \code{"1"} are interpreted as \code{TRUE},
all other strings as \code{FALSE}.
}
\examples{
dexi_bool("TRUE")
sapply(c("TRUE", "T", "1", TRUE, 1, "FALSE", "F", "0", NULL, NA, NaN), dexi_bool)

}
