% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiValues.R
\name{dexi_vector}
\alias{dexi_vector}
\title{dexi_vector}
\usage{
dexi_vector(x)
}
\arguments{
\item{x}{\code{character(1)}.}
}
\value{
Numeric vector.
}
\description{
Interpret a string, composed of \code{";"}-separated numbers, as a numeric vector.
}
\examples{
dexi_vector("1;2")
dexi_vector("1.2; 2.3")

}
