% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiValues.R
\name{distribution}
\alias{distribution}
\title{distribution}
\usage{
distribution(...)
}
\arguments{
\item{...}{Expected a comma-separated list of numeric values.}
}
\value{
An object, call it \code{obj}, such that \code{all(obj == c(...))} and
\code{class(obj) == "distribution"}.
}
\description{
Create an object as a S3 class \code{distribution}.
}
\examples{
distribution(0.1, 0.2, 0.7)

}
\seealso{
\link{DEXiR-package}, \code{\link[=set_to_distr]{set_to_distr()}}, \code{\link[=distr_to_set]{distr_to_set()}}
}
