% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiScales.R
\name{equal_scales}
\alias{equal_scales}
\title{equal_scales}
\usage{
equal_scales(scl1, scl2)
}
\arguments{
\item{scl1}{A \link{DexiScale} (or derived) object, or \code{NULL}.}

\item{scl2}{A \link{DexiScale} (or derived) object, or \code{NULL}.}
}
\value{
\code{logical(1)}.
}
\description{
Check if two scales are equal. \code{NULL} arguments, indicating undefined scales, are allowed.
Two \code{NULL} scales are considered equal.
}
