% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiUtils.R
\name{flat_text}
\alias{flat_text}
\title{flat_text}
\usage{
flat_text(value)
}
\arguments{
\item{value}{Any object that can occur as an argument of \code{\link[=c]{c()}} and \code{\link[=as.character]{as.character()}}.}
}
\value{
\code{character(1)}.
}
\description{
"Flatten" the function argument using \code{c(value)}, concatenate the elements and separate them by
a single space.
}
