% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiValues.R
\name{has_bad}
\alias{has_bad}
\title{has_bad}
\usage{
has_bad(value, scale)
}
\arguments{
\item{value}{A DEXi value.}

\item{scale}{A \link{DexiScale} or derived object.}
}
\value{
\code{logical(1)}. Whether or not \code{value_qualities(value, scale)} contains \code{"bad"}.
}
\description{
has_bad
}
