% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiValues.R
\name{has_good}
\alias{has_good}
\title{has_good}
\usage{
has_good(value, scale)
}
\arguments{
\item{value}{A DEXi value.}

\item{scale}{A \link{DexiScale} or derived object.}
}
\value{
\code{logical(1)}. Whether or not \verb{value_qualities(value, scale) contains }"good"`.
}
\description{
has_good
}
