% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiValues.R
\name{has_none}
\alias{has_none}
\title{has_none}
\usage{
has_none(value, scale)
}
\arguments{
\item{value}{A DEXi value.}

\item{scale}{A \link{DexiScale} or derived object.}
}
\value{
\code{logical(1)}. Whether or not \code{value_qualities(value, scale)} contains \code{"none"}.
}
\description{
has_none
}
