% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiValues.R
\name{has_quality}
\alias{has_quality}
\title{has_quality}
\usage{
has_quality(quality = EnumQuality, value, scale)
}
\arguments{
\item{quality}{A character string from \code{EnumQuality}.}

\item{value}{A DEXi value.}

\item{scale}{A \link{DexiScale} or derived object.}
}
\value{
\code{logical(1)}. Whether or not \code{value_qualities(value, scale)} contains \code{quality}.
}
\description{
has_quality
}
