% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiValues.R
\name{is_distribution}
\alias{is_distribution}
\title{is_distribution}
\usage{
is_distribution(value)
}
\arguments{
\item{value}{Any value or object to be checked.}
}
\value{
\code{logical(1)}. Returns \code{TRUE} if \code{value} is distribution.
}
\description{
Checks whether \code{value} is of \code{DexDistributionClass} or not.
}
\examples{
is_distribution(NULL)
is_distribution(3)
is_distribution("text")
is_distribution(c(1,2,3))
is_distribution(distribution(1,0,2))

}
