% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiUtils.R
\name{lin_map}
\alias{lin_map}
\title{lin_map}
\usage{
lin_map(x, imin, imax, omin = 0, omax = 1)
}
\arguments{
\item{x}{\code{numeric()}. Value(s) to be mapped.}

\item{imin}{\code{numeric()}. Lower bound of the input range.}

\item{imax}{\code{numeric()}. Upper bound of the input range.}

\item{omin}{\code{numeric()}. Lower bound of the output range.}

\item{omax}{\code{numeric()}. Upper bound of the output range.}
}
\value{
\code{numeric()}. Mapped value(s).
}
\description{
Map value \code{x} linearly from interval \verb{[imin:imax]} to \verb{[omax:omax]}.
}
\examples{
lin_map(2, 1, 3) # 0.5
}
