% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiFunctions.R
\name{make_args}
\alias{make_args}
\title{make_args}
\usage{
make_args(dim)
}
\arguments{
\item{dim}{A numeric vector containing upper bounds of the corresponding decision space dimensions.
For example, \code{dim = c(3, 4)} defines the space of \code{3 * 4 == 12} combinations.}
}
\value{
A list containing all possible value combinations. List elements are numeric vectors
of length equal to \code{length(dim)}.
}
\description{
Make a list of all possible combinations of values in a decision space defined by \code{dim}.
}
\examples{
make_args(c(3, 4))

}
