% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiValues.R
\name{norm_max}
\alias{norm_max}
\title{norm_max}
\usage{
norm_max(values, max = 1)
}
\arguments{
\item{values}{A numeric vector.}

\item{max}{\code{numeric(1)}.}
}
\value{
\code{values} normalized so that \code{max(result) == max}.
Returns unchanged \code{values} when \code{max(values) == 0}.
}
\description{
Normalize \code{values} so that \code{max(values) == max}.
}
\examples{
norm_max(c(0, 0.5, 0.7))

}
\seealso{
\code{\link[=norm_none]{norm_none()}}, \code{\link[=norm_sum]{norm_sum()}}
}
