% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiValues.R
\name{norm_none}
\alias{norm_none}
\title{norm_none}
\usage{
norm_none(values)
}
\arguments{
\item{values}{A numeric vector.}
}
\value{
Returns unchanged \code{values}.
}
\description{
A "do nothing" normalization function.
}
\examples{
norm_none(c(0, 0.5, 0.7))

}
\seealso{
\code{\link[=norm_max]{norm_max()}}, \code{\link[=norm_sum]{norm_sum()}}
}
