% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiValues.R
\name{norm_sum}
\alias{norm_sum}
\title{norm_sum}
\usage{
norm_sum(values, sum = 1)
}
\arguments{
\item{values}{A numeric vector.}

\item{sum}{\code{numeric(1)}.}
}
\value{
\code{values} normalized so that \code{sum(result) == sum}.
Returns unchanged \code{values} when \code{sum(values) == 0}
}
\description{
Normalize \code{values} so that \code{sum(values) == sum}.
}
\examples{
norm_sum(c(0, 0.5, 0.7))

}
\seealso{
\code{\link[=norm_none]{norm_none()}}, \code{\link[=norm_max]{norm_max()}}
}
