% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiEvaluate.R
\name{or_function}
\alias{or_function}
\title{or_function}
\usage{
or_function(method = EnumEvalMethod, or = NULL)
}
\arguments{
\item{method}{One of: \code{"set"} (default), \code{"prob"}, \code{"fuzzy"} or \code{"fuzzynorm"}.}

\item{or}{Some disjunctive aggregation function of the form \verb{function(num_vector)}, or \code{NULL}.}
}
\value{
Returns the function \code{or} if not \code{NULL}.
Otherwise, it determines the result depending on \code{method}:
\describe{
\item{\code{"set"}:}{\code{function(x) 1}}
\item{\code{"prob"}:}{\code{\link[base]{sum}}}
\item{\code{"fuzzy"}:}{\code{\link[base]{max}}}
\item{\code{"fuzzynorm"}:}{\code{\link[base]{max}}}
}
Fails with an error if the result is not an R function.
}
\description{
Determine the function to be used in the disjunctive aggregation step of \code{\link[=evaluate]{evaluate()}}.
}
\seealso{
\code{\link[DEXiR]{evaluate}}, \code{\link[=and_function]{and_function()}}.
}
