% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiAnalysis.R
\name{print_selective_explanation}
\alias{print_selective_explanation}
\title{print_selective_explanation}
\usage{
print_selective_explanation(explanation)
}
\arguments{
\item{explanation}{A list of lists,
containing selective explanation results produced by \code{\link[=selective_explanation]{selective_explanation()}}.}
}
\value{
\code{NULL}. Pretty-prints the contents of \code{explanation}.
}
\description{
A helper function for \code{\link[=selective_explanation]{selective_explanation()}}: Pretty-prints its results.
}
