% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiModels.R
\name{read_dexi}
\alias{read_dexi}
\title{read_dexi}
\usage{
read_dexi(dxi)
}
\arguments{
\item{dxi}{character(1). A \code{.dxi} file name or XML string.}
}
\value{
A \link{DexiModel} RC object.
}
\description{
\code{read_dexi()} reads a definition of a DEXi model from a \code{.dxi} file or XML string.
}
\examples{
CarDxi <- system.file("extdata", "Car.dxi", package = "DEXiR")
Car <- read_dexi(CarDxi)

}
\seealso{
\link{DexiModel}
}
