% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiValues.R
\name{rule_value}
\alias{rule_value}
\title{rule_value}
\usage{
rule_value(ch)
}
\arguments{
\item{ch}{A single character, such as \code{"3"} or \code{"Z"}.}
}
\value{
Corresponding integer value.
}
\description{
Values of decision rules are in \code{.dxi} files encoded using character strings, where each
individual character encodes some function value. The encoding is zero-based, so that
\code{"0"} represents the lowest ordinal number on the corresponding discrete scale.
\code{rule_value(char)} converts a single character to the corresponding ordinal value.
}
\examples{
rule_value("1")
rule_value("Z")

}
