% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiAttributes.R
\name{scale_of}
\alias{scale_of}
\title{scale_of}
\usage{
scale_of(obj)
}
\arguments{
\item{obj}{A \link{DexiAttribute} or \link{DexiScale}.}
}
\value{
A \link{DexiScale} associated with \code{obj}, or \code{NA} for an undefined scale.
}
\description{
scale_of
}
