% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiCharts.R
\name{transparent_colors}
\alias{transparent_colors}
\title{transparent_colors}
\usage{
transparent_colors(colors, percent = 50)
}
\arguments{
\item{colors}{A vector of color numbers or names.}

\item{percent}{Required color transparency, in the range \verb{[0:100]}.}
}
\value{
A vector of colors of the same length as \code{colors}.
}
\description{
A helper function for making \code{colors} transparent.
}
\details{
Requires installed package "grDevices".
}
\examples{
transparent_colors(c("red", "green", "blue"), 50)
# c("#FF00007F", "#00FF007F", "#0000FF7F")
}
