% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiUtils.R
\name{unique_names}
\alias{unique_names}
\title{unique_names}
\usage{
unique_names(names, reserved = c())
}
\arguments{
\item{names}{\code{character()}. Names to be converted to IDs.}

\item{reserved}{\code{character()}. Reserved names that should not be used as IDs.}
}
\value{
\code{character()}.
}
\description{
Convert \code{names} strings to ID strings that are unique and conformant
with R's syntactic rules for variable names.
}
\seealso{
\code{\link[base:make.unique]{base::make.unique()}}]
}
