% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexiValues.R
\name{values_to_str}
\alias{values_to_str}
\title{values_to_str}
\usage{
values_to_str(vals, add = 0)
}
\arguments{
\item{vals}{Numeric vector, containing ordinal values.}

\item{add}{An integer constant to be added to \code{vals} prior to conversion.}
}
\value{
A string representing DEXi's representation of ordinal values. Fails when
\code{vals + add} contains negative numbers.
}
\description{
Convert numbers to a DEXi string. Implements the reverse operation of \code{\link[=rule_values]{rule_values()}}.
}
\examples{
values_to_str(c(0, 1, 1, 2, 2, 10, 12))
values_to_str(c(1, 2, 2, 3, 3, 11, 13), -1)

}
